+user pay

+++ ms-shop/domain/src/main/kotlin/ru/md/shop/domain/product/biz/proc/ProductProcessor.kt
@@ -2,25 +2,29 @@ package ru.md.shop.domain.product.biz.proc
 
 import org.springframework.stereotype.Component
 import ru.md.base_domain.biz.proc.IBaseProcessor
-import ru.md.base_domain.biz.validate.chain.validateDeptIdAndAdminDeptLevelChain
 import ru.md.base_domain.biz.validate.validateAdminRole
 import ru.md.base_domain.biz.validate.validateAuthDeptLevel
-import ru.md.base_domain.dept.biz.validate.validateDeptId
 import ru.md.base_domain.image.biz.validate.validateImageId
 import ru.md.base_domain.biz.workers.finishOperation
 import ru.md.base_domain.biz.workers.initStatus
 import ru.md.base_domain.biz.workers.operation
+import ru.md.base_domain.dept.biz.validate.validateDeptId
+import ru.md.base_domain.dept.biz.workers.getCompanyDeptIdByAuthUser
+import ru.md.base_domain.dept.biz.workers.getCompanyIdByDeptId
 import ru.md.base_domain.dept.service.BaseDeptService
 import ru.md.base_domain.image.biz.chain.deleteS3ImageOnFailingChain
 import ru.md.base_domain.image.biz.workers.addImageToS3
 import ru.md.base_domain.image.biz.workers.deleteBaseImageFromS3
 import ru.md.base_domain.s3.repo.BaseS3Repository
+import ru.md.base_domain.user.biz.workers.findOwnerRole
 import ru.md.base_domain.user.biz.workers.getAuthUserAndVerifyEmail
 import ru.md.base_domain.user.service.BaseUserService
 import ru.md.cor.ICorChainDsl
+import ru.md.cor.chain
 import ru.md.cor.rootChain
 import ru.md.cor.worker
-import ru.md.shop.domain.product.biz.validate.validateProductId
+import ru.md.shop.domain.product.biz.validate.validateProductCount
+import ru.md.shop.domain.base.biz.validate.validateProductId
 import ru.md.shop.domain.product.biz.validate.validateProductName
 import ru.md.shop.domain.product.biz.validate.validateProductPrice
 import ru.md.shop.domain.product.biz.workers.*
@@ -50,7 +54,9 @@ class ProductProcessor(
 
 			operation("Создать приз", ProductCommand.CREATE) {
 				validateAndTrimProductFields()
-				validateDeptIdAndAdminDeptLevelChain()
+				getAuthUserAndVerifyEmail("Проверка авторизованного пользователя по authId")
+				validateAdminRole("Проверка наличия прав Администратора")
+				findCompanyDeptIdByOwnerOrAuthUserChain()
 				createProduct("Создаем приз")
 			}
 
@@ -65,17 +71,14 @@ class ProductProcessor(
 				getProductDetailsById("Получаем приз")
 			}
 
-			operation("Получить призы в отделе", ProductCommand.GET_BY_DEPT) {
-				validateDeptId("Проверяем deptId")
-				getProductsByDept("Получаем призы отдела")
+			operation("Получить призы в компании", ProductCommand.GET_BY_COMPANY) {
+				getAuthUserAndVerifyEmail("Проверка авторизованного пользователя по authId")
+				findCompanyDeptIdByOwnerOrAuthUserChain()
+				getProductsByDept("Получаем призы компании")
 			}
 
 			operation("Удалить приз", ProductCommand.DELETE) {
-				validateProductId("Проверяем productId")
-				getProductDetailsById("Получаем приз")
-				getAuthUserAndVerifyEmail("Проверка авторизованного пользователя по authId")
-				worker("Получаем deptId") { deptId = productDetails.product.deptId }
-				validateAuthDeptLevel("Проверка доступа к отделу")
+				validateProductIdAndAccessToProductChain()
 				deleteProduct("Удаляем приз")
 			}
 
@@ -101,6 +104,20 @@ class ProductProcessor(
 			finishOperation()
 		}.build()
 
+		private fun ICorChainDsl<ProductContext>.findCompanyDeptIdByOwnerOrAuthUserChain() {
+			findOwnerRole("Проверка наличия роли Владельца")
+			chain {
+				on { isAuthUserHasOwnerRole }
+				validateDeptId("Проверка deptId")
+				validateAuthDeptLevel("Проверка доступа к отделу")
+				getCompanyIdByDeptId("Получаем deptId Компании")
+			}
+			chain {
+				on { !isAuthUserHasOwnerRole }
+				getCompanyDeptIdByAuthUser("Получаем deptId Компании")
+			}
+		}
+
 		private fun ICorChainDsl<ProductContext>.validateProductIdAndAccessToProductChain() {
 			validateProductId("Проверяем productId")
 			getAuthUserAndVerifyEmail("Проверка авторизованного пользователя по authId")
@@ -113,6 +130,7 @@ class ProductProcessor(
 			validateProductName("Проверяем название")
 			validateProductPrice("Проверяем цену")
 			trimFieldProductDetails("Очищаем поля")
+			validateProductCount("Проверяем количество")
 		}
 	}
 }
